refactor: convert file to ESM and use vitest
- convert file from cjs to ESM
- replace ava with vitest
